requirements.txt
defusedxml==0.4.1
Django==1.10.4
django-pyodbc-azure==1.10.0.1
djangorestframework==3.5.3
oauthlib==2.0.1
Pillow==3.4.2
PyJWT==1.4.2
pyodbc==3.1.1
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.12.3
requests-oauthlib==0.7.0
six==1.10.0
